createNewAd promise api
